+++ tmdblibrary/src/main/java/com/b12kab/tmdblibrary/entities/BaseMember.java
@@ -26,4 +26,45 @@ public abstract class BaseMember {
 
     public String profile_path;
 
+    private Integer gender;
+
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getCredit_id() {
+        return credit_id;
+    }
+
+    public void setCredit_id(String credit_id) {
+        this.credit_id = credit_id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getProfile_path() {
+        return profile_path;
+    }
+
+    public void setProfile_path(String profile_path) {
+        this.profile_path = profile_path;
+    }
+
+    public Integer getGender() {
+        return gender;
+    }
+
+    public void setGender(Integer gender) {
+        this.gender = gender;
+    }
 }