+++ tmdblibrary/src/main/java/com/b12kab/tmdblibrary/services/CollectionService.java
@@ -29,6 +29,7 @@ public interface CollectionService {
      * @param tmdbId TMDb id.
      * @param language <em>Optional.</em> ISO 639-1 code.
      * @param appendToResponse <em>Optional.</em> extra requests to append to the result.
+     * @return Collection
      */
     @GET("/collection/{id}")
     Collection summary(@Path("id") int tmdbId, @Query("language") String language, @Query("append_to_response") AppendToResponse appendToResponse);
@@ -38,6 +39,7 @@ public interface CollectionService {
      *
      * @param tmdbId TMDb id.
      * @param language <em>Optional.</em> ISO 639-1 code.
+     * @return Images
      */
     @GET("/collection/{id}/images")
     Images images(@Path("id") int tmdbId, @Query("language") String language);